Creating & Joining Tables
Here, we will create tables to hold the data from the student_information.csv and the student_predictors.csv. Once these tables are created, we will join these tables using a left join on the ID column and export the table as a csv file, ready to be analyzed through the machine learning model.